Rename new Formatters->Translators and minor cleanup
